remove git anomalies
Somehow, code from another branched that I explicitly removed in this branch ended back up in this branch, and sometimes duplicated it's self. I believe this is caused by switching branches, but I'm not exactly sure on the specifics since it shouldn't be bring back snippets I manually removed myself.